feat: Locale switcher to footer
